add methods and split code into components
