+++ src/api/todos.ts
@@ -1,8 +1,22 @@
 import { Todo } from '../types/Todo';
 import { client } from '../utils/fetchClient';
 
-export const getTodos = (userId: number) => {
+export const getTodos = (userId: number | undefined) => {
   return client.get<Todo[]>(`/todos?userId=${userId}`);
 };
 
-// Add more methods here
+export const addTodo = (userId: number | undefined, title: string) => {
+  return client.post<Todo>('/todos', {
+    userId,
+    completed: false,
+    title,
+  });
+};
+
+export const updatingTodo = (todoId: number, completed: boolean) => {
+  return client.patch(`/todos/${todoId}`, { completed });
+};
+
+export const deleteTodo = (todoId:number) => {
+  return client.delete(`/todos/${todoId}`);
+};
