+++ src/components/ErrorNotification/ErrorNotification.tsx
@@ -1,0 +1,18 @@
+/* eslint-disable jsx-a11y/control-has-associated-label */
+import React from 'react';
+
+export const ErrorNotification: React.FC = () => {
+  return (
+    <div
+      data-cy="ErrorNotification"
+      className="notification is-danger is-light has-text-weight-normal"
+    >
+      <button data-cy="HideErrorButton" type="button" className="delete" />
+      Unable to add a todo
+      <br />
+      Unable to delete a todo
+      <br />
+      Unable to update a todo
+    </div>
+  );
+};
